Split tests into their own files
As a step forward towards moving them to the actual Yocto layers.

Change-Type: patch
